Second method

+++ New .cpp
@@ -6,7 +6,7 @@ int main()
 {
     double x, y, z, a, b;
     scanf_s("%lf %lf %lf", &x, &y, &z);
-    a = pow(x, 3) * (pow(tan(pow(x + y, 2)), 2) + z / (sqrt(x + y) - 2 * tan(0.436)));
-    b = (y * pow(x, 2) - z) / (exp(y * x) - 1) + (pow(5 - pow(y, 3), 1 / 3) / (abs(x - y) / sqrt(2 * x)));
+	a = static_cast <double>(pow(x, 3)) * (pow(tan(pow(x + y, 2)), 2) + z / (sqrt(x + y) - 2 * tan(0.436)));
+	b = static_cast <double>(y * pow(x, 2) - z) / (exp(y * x) - 1) + (pow(5 - pow(y, 3), 1 / 3) / (abs(x - y) / sqrt(2 * x)));
     printf("a=%lf\n b=%lf", a, b);
 }
